app/server/requirements.txt
acoular==24.10
alembic==1.15.2
annotated-types==0.7.0
anyio==4.9.0
appnope==0.1.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
attrs==24.2.0
autopep8==2.3.2
babel==2.16.0
beautifulsoup4==4.12.3
bleach==6.1.0
blosc2==2.7.1
boto3==1.40.26
botocore==1.40.26
branca==0.8.1
build==1.2.2.post1
CacheControl==0.14.3
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.3.2
cleo==2.1.0
click==8.1.7
colorednoise==2.2.0
comm==0.2.2
contourpy==1.2.1
crashtest==0.4.1
cryptography==44.0.2
cycler==0.12.1
dacite==1.8.1
debugpy==1.8.5
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.9
distro==1.9.0
dulwich==0.22.8
executing==2.0.1
face_recognition_models==0.3.0
fastapi==0.115.12
fastjsonschema==2.20.0
filelock==3.18.0
findpython==0.6.3
folium==0.19.2
fonttools==4.53.1
fqdn==1.5.1
geopandas==1.0.1
h11==0.14.0
htmlmin==0.1.12
httpcore==1.0.5
httpx==0.27.2
idna==3.8
ImageHash==4.3.1
imageio==2.36.1
installer==0.7.0
ipykernel==6.29.5
ipympl==0.9.4
ipython==8.26.0
ipython-genutils==0.2.0
ipywidgets==8.1.5
isoduration==20.11.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jedi==0.19.1
Jinja2==3.1.4
jiter==0.11.1
jmespath==1.0.1
joblib==1.4.2
json5==0.9.25
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-highlight-selected-word==0.2.0
jupyter-lsp==2.2.5
jupyter_client==8.6.2
jupyter_contrib_core==0.4.2
jupyter_contrib_nbextensions==0.7.0
jupyter_core==5.7.2
jupyter_nbextensions_configurator==0.6.4
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
jwt==1.3.1
keyring==25.6.0
kiwisolver==1.4.5
langchain-core==1.0.2
langchain-openai==1.0.1
langgraph==1.0.2
langgraph-checkpoint==3.0.0
langgraph-prebuilt==1.0.2
langgraph-sdk==0.2.9
langsmith==0.4.38
lazy_loader==0.4
llvmlite==0.43.0
lxml==5.3.0
m2cgen==0.10.0
Mako==1.3.2
Markdown==3.6
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.7
mistune==3.0.2
more-itertools==10.7.0
mpmath==1.3.0
msgpack==1.1.0
multimethod==1.12
nanoid==2.0.0
narwhals==1.36.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
ndindex==1.9.2
nest-asyncio==1.6.0
networkx==3.3
nltk==3.9.1
notebook==7.2.2
notebook_shim==0.2.4
numba==0.60.0
numexpr==2.10.1
numpy==1.26.4
openai==2.6.1
opencv-python==4.10.0.84
orjson==3.11.4
ormsgpack==1.11.0
overrides==7.7.0
packaging==24.1
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
patsy==0.5.6
pbs-installer==2025.5.17
pexpect==4.9.0
phik==0.12.4
pillow==10.4.0
pkginfo==1.12.1.2
platformdirs==4.2.2
plotly==6.0.1
poetry==2.1.3
poetry-core==2.1.3
prometheus_client==0.20.0
prompt_toolkit==3.0.47
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pycodestyle==2.13.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pydotplus==2.0.2
pygame==2.6.1
Pygments==2.18.0
PyMuPDF==1.26.4
pyogrio==0.9.0
pyparsing==3.1.4
pyplnoise==1.4
pyproj==3.6.1
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.1
python-json-logger==2.0.7
python-multipart==0.0.20
pytz==2024.1
PyWavelets==1.7.0
PyYAML==6.0.2
pyzmq==26.2.0
RapidFuzz==3.13.0
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.20.0
s3transfer==0.13.1
scikit-image==0.24.0
scikit-learn==1.5.2
scipy==1.13.1
seaborn==0.13.2
Send2Trash==1.8.3
setuptools==69.2.0
shapely==2.0.6
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.40
sqlmodel==0.0.24
sse-starlette==2.3.5
stack-data==0.6.3
starlette==0.46.1
statsmodels==0.14.2
sympy==1.13.3
tables==3.10.1
tenacity==9.1.2
terminado==0.18.1
threadpoolctl==3.5.0
tifffile==2024.9.20
tiktoken==0.12.0
tinycss2==1.3.0
tomlkit==0.13.2
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
traits==6.4.3
trove-classifiers==2025.5.9.12
typeguard==4.3.0
types-python-dateutil==2.9.0.20240906
typing_extensions==4.12.2
tzdata==2024.1
uri-template==1.3.0
urllib3==2.2.2
uvicorn==0.34.0
virtualenv==20.31.2
visions==0.7.6
wcwidth==0.2.13
webcolors==24.8.0
webencodings==0.5.1
websocket-client==1.8.0
widgetsnbextension==4.0.13
wordcloud==1.9.3
xattr==1.1.4
xxhash==3.6.0
xyzservices==2024.9.0
ydata-profiling==4.9.0
zstandard==0.23.0
